+++ armstrong/utils/backends/__init__.py
@@ -1,1 +1,4 @@
-from .base import DID_NOT_HANDLE, GenericBackend
+from .base import GenericBackend, BackendDidNotHandle
+
+# DEPRECATED: To be removed in Backends 2.0
+from .base import DID_NOT_HANDLE