+++ tests/support/backends.py
@@ -1,6 +1,6 @@
 import fudge
 
-from armstrong.utils.backends.base import DID_NOT_HANDLE
+from armstrong.utils.backends.base import BackendDidNotHandle
 
 
 TestArgs = fudge.Fake().expects('__init__').with_args('arg1', kw=1)
@@ -8,7 +8,7 @@ TestArgs = fudge.Fake().expects('__init__').with_args('arg1', kw=1)
 
 class Skip(object):
     func = fudge.Fake().is_callable()\
-        .with_args('arg1', kw=1).returns(DID_NOT_HANDLE)
+        .with_args('arg1', kw=1).raises(BackendDidNotHandle)
 
 
 class UseThisOne(object):