+++ CHANGES.rst
@@ -4,6 +4,9 @@ CHANGES
 1.1.0 (unreleased)
 ------------------
 
+- **DEPRECATION:** deprecate DID_NOT_HANDLE. Raising a ``BackendDidNotHandle``
+  exception is now the preferred way to skip when using multiple backends.
+
 - Expanded documentation with examples
 
 - Use Setuptools and pkg_resources for namespacing